fix: Move the Empty state
